requirements.txt
gradio
google-generativeai
PyPDF2
Pillow
pytesseract
speechrecognition
datasets
python-dotenv
